requirements.txt
absl-py==2.1.0
albumentations==1.3.1
astunparse==1.6.3
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
gdown==4.7.3
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.8.0
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
joblib==1.3.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
networkx==2.6.3
nibabel==4.0.2
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==23.2
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydicom==2.4.4
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==68.0.0
six==1.16.0
soupsieve==2.4.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
tqdm==4.66.2
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wheel==0.41.2
wrapt==1.16.0
zipp==3.15.0
PyQt5==5.15.10